Add tooltip/popup features
The PR implements tooltip to Markers and tooltip/popup to Polylines
I created two concerns for those purposes
