+++ src/Concerns/HasTooltip.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Webbingbrasil\FilamentMaps\Concerns;
+
+trait HasTooltip
+{
+    protected string | Closure | null $tooltip = null;
+
+    public function tooltip(string| Closure | null $tooltip): static
+    {
+        $this->tooltip = $tooltip;
+
+        return $this;
+    }
+
+    public function getTooltip(): ?string
+    {
+        return $this->evaluate($this->tooltip);
+    }
+}
